+++ app/guolaiwan-app2.0/guolaiwan-app-web Maven Webapp/src/main/java/com/guolaiwan/app/web/admin/vo/LiveGiftVO.java
@@ -8,7 +8,7 @@ public class LiveGiftVO extends AbstractBaseVO<LiveGiftVO,LiveGiftPO>{
 
 	private String name;
 	
-	private long price;
+	private int price;
 	
 	private long picId;
 	
@@ -24,11 +24,13 @@ public class LiveGiftVO extends AbstractBaseVO<LiveGiftVO,LiveGiftPO>{
 		this.name = name;
 	}
 
-	public long getPrice() {
+	
+
+	public int getPrice() {
 		return price;
 	}
 
-	public void setPrice(long price) {
+	public void setPrice(int price) {
 		this.price = price;
 	}
